Fix Comments & Killswitch
Comment blocks now comment out only the declaration, as it is no longer able to be triggered despite the code existing below.

The Killswitch has been reduced to a single line for efficiency.

The Message Box detailing hotkeys has been moved to the top of the script to prevent not being reached upon initialization.